Split websocket implementation into client and server files
